+++ cypress-shared/support/common/testcase.js
@@ -307,7 +307,15 @@ export function syncCheckoutUICustom() {
   it(`In ${prefix} - Sync Checkout UI Custom`, updateRetry(2), () => {
     cy.visit('admin/app/vtex-checkout-ui-custom/')
     cy.contains('Publish', { timeout: 25000 }).should('be.visible').click()
+    cy.getVtexItems().then((vtex) => {
+      cy.intercept('POST', `${vtex.baseUrl}/**`, (req) => {
+        if (req.body.operationName === 'GetHistory') {
+          req.continue()
+        }
+      }).as('GetHistory')
+    })
     cy.contains('History', { timeout: 35000 }).should('be.visible').click()
+    cy.wait('@GetHistory')
     cy.contains(WORKSPACE, { timeout: 15000 }).should('be.visible')
   })
 }