+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 ### Added
+- For SyncCheckout UI Custom testcase - intercept GetHistory graphql
+
+### Added
 - Shorter report on PR decorator
 - Enable quickOrderBySkuAnd51QuantityTestCase - Instead of toast message verify tooltip message